Better support for orientations
